more comments

+++ mediaire_toolbox/data_cleaner.py
@@ -24,6 +24,9 @@ class DataCleaner:
         """
         Parameters
         ----------
+        folder:
+            folder: str
+            Path of folder to be cleaned.
         max_folder_size: int
             Max folder size, delete until current size is smaller than this.
             -1 if not deleting files based on size.
@@ -84,22 +87,23 @@ class DataCleaner:
                 shutil.rmtree(folder)
             return None
         # the set of filtered filenames that match the whitelist/blacklist pattern
-        filter_set = set()
+        filtered_files_set = set()
         # the list of patterns from whitelist/blacklist
         filter_list = self.whitelist or self.blacklist
+        # for every pattern in the whitelist/blacklist, match and join results
         for f in filter_list:
-            filter_set = filter_set.union(set(fnmatch.filter(
+            filtered_files_set = filtered_files_set.union(set(fnmatch.filter(
                                           file_set, f)))
 
         if self.whitelist:
             # if it is a whitelist, the list of deleted files are the files in the folder
             # but not on the matched filenames
             delete_list = [os.path.join(folder, f) for f in 
-                           list(file_set - filter_set)]
+                           list(file_set - filtered_files_set)]
         else:
             # if it is a blacklist, the list of deleted files are the files with matched filenames
             delete_list = [os.path.join(folder, f) for f in 
-                           list(filter_set)]
+                           list(filtered_files_set)]
 
         for file_path in delete_list:
             if dry_run:
@@ -108,12 +112,18 @@ class DataCleaner:
                 default_logger.info('Removing file [%s]' % file_path)
                 os.remove(file_path)
         # remove the folder if it is empty
-        if len(os.listdir(folder)) == 0:
-            shutil.rmtree(folder)
+        if not dry_run:
+            if len(os.listdir(folder)) == 0:
+                shutil.rmtree(folder)
         return delete_list
 
     def clean_up(self, dry_run=False):
         """Cleans up folders that either take up too much space or are too old
+        
+        Returns
+        -------
+        list
+            Returns the list of folders cleanded.
         """
         removed = []
         if self.max_folder_size == -1 and self.max_data_seconds == -1:
